+++ docs/source/changelog.rst
@@ -1,6 +1,33 @@
 Changelog
 =========
 
+Version 1.0.0b1 (beta version)
+------------------------------
+
+*Release date: 2021/01/04*
+
+* changed: **requires Python 3.5+ now**
+* changed: specifying trackers is now optional with the new ``-t`` switch, thus **trackerless torrents are now
+  possible**
+* added: **bestN shortcut**! It is now possible to add the best N trackers from `ngosang/trackerslist <https://github.com/ngosang/trackerslist>`_
+  by using the new `bestN shortcut <user.html#bestn-automatically-add-the-best-trackers>`__
+* changed: use **external JSON files for configuration**, by default try to load ``.py3createtorrent.cfg``
+  from user's home directory
+* added: DHT bootstrap nodes can now be specified with the new ``--node`` switch (doing so is recommended for
+  trackerless torrents)
+* added: **webseed support** with the new ``--webseed`` switch (GetRight style, i.e. `<http://bittorrent.org/beps/bep_0019.html>`_)
+* changed: increased max piece size to 16 MiB
+* changed: show warning if piece size is not a multiple of 16 KiB
+* changed: updated the default trackers (openbt is now opentrackr, dropped publicbt, added cyberia and coppersurfer)
+* added: Pipfile and Pipfile.lock for pipenv support
+* added: README.md
+* changed: improved performance of single file torrent creation
+* refactored: switched to bencode.py module for encoding the torrent data
+* refactored: switched from optparse to argparse
+* refactored: reformatted code with yapf, using a new column limit of 120
+* refactored: added type hints to enable analysis with mypy (we use Python's typing module which was added in
+  Python 3.5, thus Python 3.5 is the new minimum version that is required)
+
 Version 0.9.7
 -------------